+++ crates/file-group/src/file_group.rs
@@ -67,6 +67,9 @@ impl FileGroup {
             Input::ProjectGlob(glob) | Input::WorkspaceGlob(glob) => {
                 self.globs.push(glob.to_workspace_relative(project_source));
             }
+            Input::ProjectSources(_) => {
+                // Skip
+            }
         };
 
         Ok(())