+++ crates/task-expander/src/token_expander.rs
@@ -300,6 +300,9 @@ impl<'graph> TokenExpander<'graph> {
                         .globs_for_input
                         .insert(glob, TaskGlobInput { cache: inner.cache });
                 }
+                Input::ProjectSources(_) => {
+                    // Skip
+                }
             };
         }